First try to dynamically create instances
I have put there the code that allowed me to create an instance last time.
Let's assume it works (it did last week) untile we hear back from Sanjay.

Added a get_config function to load the configuration file from a json.
We might want to use a .ini like file later on, but json is quicker :)
For the moment we need something like:
{"region": "us-west-2", "aws_access_key_id": "XXX", "aws_secret_access_key": "YYY"}
and this has to be passed from the command line.

Fixed some docstrings (made them PEP-0008 compliant)